WIP: Introducing the initial commit of BAM Modder.
What is BAM Modder:
After read many feedbacks and people blaming for additional mod, like launchers, audio mods etc. i decided to create this.
BAM Modder is a simple menu that display in lists some mods that will be downloaded to /sdcard/jellybam/ and ready to be flashed
via recovery. (Maybe one day, during another boring time i'll figure out a way to auto-install them)

The fragments, as well as the general UI will be injected into settings app. Otherwise, the single mod fragment with info and download
will be done by a separate dedicated app that will be installed as prebuilt (i'll keep it closed source for the moment)

Once again, this is an initial part of the work... the layout and the fragments needs to be created so please, if you build jellybam
from sources wait a few day for do it and release an incomplete version.

For the moment, the bammodder will provide the following mods category: Launchers, Audio, Graphics and Miscellaneus.

Feel free to give any suggestions at giovanni.rocca.90@gmail.com