Add Date to Status Bar clock (2/2) This commit adds the ability to include the date in the Status Bar Clock, it will add the following: 13 April or short version 13 Apr. The motivation for adding this feature was user requests, in particular xda member JokerAce who provided details of a SystemUI hack he was using to implement this post build.
If enabled the date resides to the left of the time display, and if the day name has been enabled the date will be to the right of the day, so with everything on it would resemble: Sat 13 April 3:50PM (or with the short date).

Perference options added to Rom Control as the other settings for the Status Bar Clock are also found there, it made sense to include this feature there also for a seamless user experience.

Compiled and tested with no problems, I tried all different combinations all of which produced the desired effects.

Screenshot of menu which also shows the enabled date:
http://i45.tinypic.com/2lu9ri9.jpg

Change-Id: Ifa1c1f8e72fc0b2a478598349d9794e94d505c31

Signed-off-by: igio90 <giovanni.rocca.90@gmail.com>